feat: complement system unit test
