+++ packages/system/tests/system-in-node.test.tsx
@@ -1,0 +1,32 @@
+/**
+ * @jest-environment node
+ */
+import { raf, caf, throttleByRaf } from "../src"
+
+test(`Test raf in node environment`, async () => {
+  const fn = jest.fn()
+  jest.useFakeTimers()
+  const timer = raf(fn)
+  jest.runAllTimers()
+  caf(timer)
+
+  expect(fn).toBeCalled()
+})
+
+test(`Test throttleByRaf`, async () => {
+  jest.useFakeTimers()
+  let count = 0
+  const fn = jest.fn(() => (count += 1))
+  const fnThrottle = throttleByRaf(fn)
+
+  const interval = setInterval(() => {
+    fnThrottle()
+  }, 16)
+  setTimeout(() => {
+    clearInterval(interval)
+    fnThrottle.cancel()
+  }, 1000)
+
+  jest.runAllTimers()
+  expect(count).toBeGreaterThanOrEqual(60)
+})
